Move and restructure election components
