+++ frontend/src/api/election/index.ts
@@ -1,6 +1,4 @@
-export * from "./ElectionListProvider";
 export * from "./ElectionProvider";
 export * from "./ElectionStatusProvider";
 export * from "./useElection";
-export * from "./useElectionList";
 export * from "./useElectionStatus";
