+++ frontend/src/app/ElectionLayout.tsx
