+++ frontend/src/components/form/data_entry/state/reducer.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "vitest";
 
-import { mockElection } from "@/components/election/status/mockData";
+import { electionMockData } from "@/testing/api-mocks";
 
 import { ApiResponseStatus, Election, PollingStationResults } from "@kiesraad/api";
 
@@ -8,7 +8,7 @@ import dataEntryReducer, { getInitialState as _getInitialState } from "./reducer
 import { DataEntryAction, DataEntryState } from "./types";
 
 function getInitialState(): DataEntryState {
-  return _getInitialState(mockElection, 1, 1);
+  return _getInitialState(electionMockData, 1, 1);
 }
 
 export function _getInitialValues(
@@ -52,7 +52,7 @@ export function _getInitialValues(
 }
 
 function getInitialValues() {
-  return _getInitialValues(mockElection);
+  return _getInitialValues(electionMockData);
 }
 
 test("should handle DATA_ENTRY_CLAIMED with client_state", () => {
