+++ frontend/src/features/data_entry/components/DataEntryHomePage.tsx
@@ -1,13 +1,14 @@
 import { useEffect } from "react";
 import { useLocation, useNavigate } from "react-router";
 
-import { ElectionProgress } from "@/components/election/ElectionProgress";
 import { Footer } from "@/components/footer/Footer";
 import { PollingStationChoiceForm } from "@/components/form/data_entry/polling_station_choice/PollingStationChoiceForm";
 
 import { DEFAULT_CANCEL_REASON, useElection, useElectionStatus } from "@kiesraad/api";
 import { t } from "@kiesraad/i18n";
 import { Alert, PageTitle } from "@kiesraad/ui";
+
+import { ElectionProgress } from "./ElectionProgress";
 
 export function DataEntryHomePage() {
   const navigate = useNavigate();
