+++ frontend/src/features/election_overview/hooks/useElectionList.ts
