+++ frontend/src/features/election_overview/hooks/useElectionListRequest.ts
@@ -1,4 +1,4 @@
-import { ELECTION_LIST_REQUEST_PATH, ElectionListResponse, useInitialApiGetWithErrors } from "@kiesraad/api";
+import { ELECTION_LIST_REQUEST_PATH, ElectionListResponse, useInitialApiGetWithErrors } from "@/api";
 
 export function useElectionListRequest() {
   const path: ELECTION_LIST_REQUEST_PATH = "/api/elections";
