+++ frontend/src/features/election_status/components/ElectionStatus.module.css
