+++ frontend/src/features/election_status/components/ElectionStatusPage.tsx
@@ -1,12 +1,12 @@
 import { useNavigate } from "react-router";
 
-import { HeaderElectionStatusWithIcon } from "@/components/election/ElectionStatusWithIcon";
-import { ElectionStatus } from "@/components/election/status/ElectionStatus";
+import { useElection, useElectionStatus } from "@/api";
+import { HeaderElectionStatusWithIcon } from "@/components/election_status_with_icon/ElectionStatusWithIcon";
 import { Footer } from "@/components/footer/Footer";
+import { Alert, Button, PageTitle } from "@/components/ui";
+import { t } from "@/lib/i18n";
 
-import { useElection, useElectionStatus } from "@kiesraad/api";
-import { t } from "@kiesraad/i18n";
-import { Alert, Button, PageTitle } from "@kiesraad/ui";
+import { ElectionStatus } from "./ElectionStatus";
 
 export function ElectionStatusPage() {
   const navigate = useNavigate();
